+++ client/src/components/LoginForm.vue
@@ -8,7 +8,7 @@
   import isEmail from 'validator/es/lib/isEmail';
   import { reactive } from 'vue';
   import { useRouter } from 'vue-router';
-  import type { FormFieldState } from './types';
+  import type { FormFieldState } from '../types';
 
   type LoginFormState = {
     fields: Record<'email' | 'password', FormFieldState>;
