--- client/src/components/types.ts
@@ -1,4 +1,0 @@
-export type FormFieldState = {
-  value: string;
-  errorMessage: string;
-};
